+++ Evergreen/Apps/Get-ESETServerSecurity.ps1
@@ -1,4 +1,4 @@
-function Get-ESETSecureAuthentication {
+function Get-ESETServerSecurity {
     <#
         .NOTES
             Site: https://stealthpuppy.com